Enable multiple console proxies when required in deployments
When Nova is deployed with a mix of x86 and arm systems
(for example), it may be necessary to deploy both 'novnc' and
'serialconsole' proxy services on the same host in order to
service the mixed compute estate.

This patch introduces a list which defines the required proxy
console types.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-os_nova/+/890521

Change-Id: I5ed49878c192516a504a4a77902271214800c5b8
(cherry picked from commit 66d0dc7626656457b259a142c06bf8428a00c0b5)