Update cinder SHA to pull in 3PAR concurrency issues
This commit pulls in the cinder patches to fix various HPE 3PAR cinder
driver concurrency issues that occur when multiple concurrent volume
creation and deletion requests are executed.

The current set of patches improves the stability to the level that it
is safe to run up to 10 concurrent operations. More concurrencies than
this value can result in long-term instability.

For now, until we have a more fundamental solution to the big mess that
the 3PAR cinder driver is, use this patch and ensure that volume
operation concurrency is maintained at a relatively low level.

Upstream-Status: Inappropriate
Signed-off-by: Stephanos Ioannidis <root@stephanos.io>